TESE_DOUTORAMENTO/09_IMPLEMENTACAO_TECNICA/requirements.txt
# Base de Dados
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Análise de Dados
pandas==2.1.4
numpy==1.26.2

# Parsing de PDFs
pdfplumber==0.10.3
PyPDF2==3.0.1
tabula-py==2.9.0

# Validação
pydantic==2.5.2
python-dotenv==1.0.0

# Visualização (para análises rápidas)
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Machine Learning (se precisar já conectar)
scikit-learn==1.3.2
xgboost==2.0.3

# Utilities
tqdm==4.66.1
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Logging
loguru==0.7.2
